[Update]BE/controller,service/controller return value,service exception update 드디어 끝났다...

+++ Back/src/main/java/com/ssafy/howdoilook/domain/follow/service/FollowService.java
@@ -9,6 +9,7 @@ import com.ssafy.howdoilook.domain.follow.repository.FollowReposiroty;
 import com.ssafy.howdoilook.domain.user.entity.User;
 import com.ssafy.howdoilook.domain.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
@@ -29,9 +30,9 @@ public class FollowService {
     @Transactional
     public Long saveFollow(FollowSaveRequestDto followSaveRequestDto){
         User follower = userRepository.findById(followSaveRequestDto.getFollowerId())
-                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 follower입니다."));
+                .orElseThrow(() -> new EmptyResultDataAccessException("존재하지 않는 follower입니다.",1));
         User followee = userRepository.findById(followSaveRequestDto.getFolloweeId())
-                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 followee입니다."));
+                .orElseThrow(() -> new EmptyResultDataAccessException("존재하지 않는 followee입니다.",1));
         //넘어온 팔로우,팔로워 값으로 이미 데이터가 있는지 확인하는 작업
         //null값이 아니면 이미 데이터가 있는 것이다.
         Follow findFollow = followReposiroty.findFollowIdByFollowerAndFollowee(follower.getId(), followee.getId());
@@ -55,12 +56,15 @@ public class FollowService {
     public void deleteFollow(FollowDeleteRequestDto followDeleteRequestDto){
         Follow findFollow = followReposiroty.findFollowIdByFollowerAndFollowee(
                 followDeleteRequestDto.getFollowerId(), followDeleteRequestDto.getFolloweeId());
-        followReposiroty.deleteById(findFollow.getId());
+        //팔로우 관계가 있을 때
+        if (findFollow != null) {
+            followReposiroty.deleteById(findFollow.getId());
+        }
     }
     //내가 팔로우 하는 사람 리스트 반환
     public Page<FolloweeResponseDto> selectFolloweeList(Long userId,Pageable page){
         User findUser = userRepository.findById(userId).orElseThrow(
-                ()->new IllegalArgumentException("존재하지 않는 User 입니다."));
+                ()->new EmptyResultDataAccessException("존재하지 않는 User입니다.",1));
 
         Page<Follow> followeeByUserId = followReposiroty.findFolloweeByUserId(userId, page);
         List<Follow> content = followeeByUserId.getContent();
@@ -85,7 +89,7 @@ public class FollowService {
     public Page<FollowerResponseDto> selectFollowerList(Long userId, Pageable page){
 
         User findUser = userRepository.findById(userId).orElseThrow(
-                ()->new IllegalArgumentException("존재하지 않는 User 입니다."));
+                ()->new EmptyResultDataAccessException("존재하지 않는 User 입니다.",1));
 
         Page<Follow> followerByUserId = followReposiroty.findFollowerByUserId(userId, page);
         List<Follow> content = followerByUserId.getContent();
